bsc: split the case-building transformations out of AOpt
AOpt exports aExpandDynSel and aInsertCaseDef for the Bluesim backend, and
SimPackageOpt imports exactly those two.  AOpt also runs optAndOrDefs, whose
and/or redundancy check goes through SAT to AExpr2STP and AExpr2Yices, so
importing AOpt for the two transformations linked STP and Yices as well.
Neither transformation asks a solver anything: each rewrites one expression
tree into another and takes the def lookup it needs as an argument.

AOptCase now holds them, together with the helpers only they use --
expandVarRef, collIf/collIfPrim, getConsts, isInt, rmDupsInCasePairs.  AOpt
imports isInt and rmDupsInCasePairs back, since it builds case arms of its
own, and keeps the passes that do need a solver.  This is code motion: no
definition changed.

The reach of the old edge was wider than the solvers.  Measured on the BIR
exporter, whose only door into AOpt was SimPackageOpt, the module closure
drops from 125 to 102: out go SAT, STP, STPFFI, Yices, YicesFFI, AExpr2STP
and AExpr2Yices, and with them the tail AOpt pulled behind it -- AConv,
AExpand, BDD, BExpr, BoolExp, BoolOpt, Changed, IInline, IInlineUtil,
Intervals, IPrims, ISyntaxSubst, ISyntaxUtil, ITransform, SignalNaming.
Neither vendored solver is built or linked for that program now.

Verified with bsc.codegen: 641 expected passes, 18 expected failures, 0
unexpected.  That is the golden-file suite over generated Verilog and Bluesim
C++, which is the output these two transformations shape.  bsc.bluesim shows
34 failures on this box that this change does not touch: 23 SystemC cases
where SYSTEMC_INC is unset in the test environment, and 11 FST cases where
Bluesim writes correct .fst files and the harness's own fst2vcd conversion
step does not run -- confirmed by converting one by hand.

The split stands on its own as a compiler change and does not depend on the
trs work that motivated it.
